# backend/core-api/src/models/index.js
const User = require("./User");
const Group = require("./Group");
const GroupMember = require("./GroupMember");
const Password = require("./Password");
const PasswordShare = require("./PasswordShare");
const UserEncryptionKey = require("./UserEncryptionKey");

module.exports = {
  User,
  Group,
  GroupMember,
  Password,
  PasswordShare,
  UserEncryptionKey,
};